[Refactoring]: 이전에 프로퍼티래퍼를 사용하지 않은 것을 수정하였습니다.
